refactor: 폴더 구조 및 변수 네이밍 가독성 개선
